+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-### Ribbon V2.9.2, RibbonTools V1.3.4
+### Ribbon V2.10.0, RibbonTools V1.3.4
+#### Changed (Ribbon)
+- new Ribbon event ViewDestroy (Example: you can use it for saving the settings)
+- new Properties (Tag, CommandType) for all Ribbon Controls (RibbonButton, ...)
+- Interface IRibbonControl is now public
+- some bugfixes
 
 #### Changed (RibbonTools)
 - Some Setting dialog text and tooltips modified.